Phase-1 models/Transcription_translation/Translation with phi llm/requirements.txt
arrow==1.3.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
circuitbreaker==2.1.3
click==8.0.4
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
decorator==5.2.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oci==2.154.1
oci-cli==3.58.1
openai-whisper==20240930
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt-toolkit==3.0.43
protobuf==6.30.2
PyAudio==0.2.14
pycparser==2.22
pyOpenSSL==24.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.4.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
charset-normalizer==3.4.2
diskcache==5.6.3
gpt4all==2.8.2
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.2.81
MarkupSafe==3.0.2
tqdm==4.67.1
